styling van update functie verbeterd

+++ view/location/locationDetails.php
@@ -61,7 +61,7 @@ foreach($result as $location){ ?>
                     </div>
 
                     <div class="<?= $show_class?>">
-                        <a href="<?= URL ?>Location/deleteLocation?id=<?= $location['location_id']?>" class="btn btn-danger w-50 ">Verwijder</a>
+                        <a href="<?= URL ?>Location/deleteLocation?id=<?= $location['location_id']?>" class="btn btn-danger w-50 " onclick="return confirm('Weet u zeker dat u deze locatie wilt verwijderen')">Verwijder</a>
                         <a href="<?= URL ?>Location/updateLocation?id=<?= $location['location_id']?>" class="btn btn-warning w-50">Update</a>
